Revert adding AuthRejection to common
